[Core] IntRange.RandomWithin now correctly returns an int instead of a float

+++ _Demigiant.Libraries/Core/DemiLib/Structs (no namespace)/IntRange.cs
@@ -29,7 +29,7 @@ namespace DG.DemiLib
         /// <summary>
         /// Returns a random value within this range (min/max included)
         /// </summary>
-        public float RandomWithin()
+        public int RandomWithin()
         {
             return UnityEngine.Random.Range(min, max + 1);
         }
